+++ client/src/test/unit/lib/parse/search/timelineParser.spec.ts
@@ -87,7 +87,7 @@ describe('timelineParser', () => {
         mockJsonCriteria,
       )
       expect(formattedYear).toEqual(
-        '{"AND":[{"producedBy":{"id":"https://endpoint.yale.edu/data/person/783e7e6f-6863-4978-8aa3-9e6cd8cd8e83"}},{"producedDate":"1981","_comp":"<"},{"producedDate":"1979","_comp":">"}]}',
+        '{"AND":[{"producedBy":{"id":"https://endpoint.yale.edu/data/person/783e7e6f-6863-4978-8aa3-9e6cd8cd8e83"}},{"producedDate":"1981-01-01T00:00:00.000Z","_comp":"<="},{"producedDate":"1980-01-01T00:00:00.000Z","_comp":">="}]}',
       )
     })
   })